+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added SESSION_EXPIRES_IN_SECONDS config to myems-api
 - added virtual power plant to database, myems-api and myems-admin
 - added Russian language to myems-admin
+- added higher limit and lower limit of point to myems-api and myems-admin
 - 
 ### Changed
 - changed fdd rule template in myems-admin